backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.5.2
pyodbc>=5.1.0,<5.4
alembic==1.13.3
python-dotenv==1.0.1
loguru==0.7.2
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
jinja2
pypinyin
python-multipart
